+++ tests/timer-episode-range.test.js
@@ -45,6 +45,7 @@ function realSumEpisodeRuntimes(seasonDetails, episodeStart, episodeEnd, fallbac
 }
 
 jest.unstable_mockModule('../src/services/tmdbService.js', () => ({
+  getPosterUrl: jest.fn(() => null),
   searchMovies: mockSearchMovies,
   searchTVShows: mockSearchTVShows,
   getMovieDetails: mockGetMovieDetails,